% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcw2XML.R
\name{dcw2XML}
\alias{dcw2XML}
\title{dcw2XML - Convert Darwin Core format data frame to XML}
\usage{
dcw2XML(dat = NA)
}
\arguments{
\item{dat}{- Darwin Core Data frame to convert to XML}
}
\value{
Simple Darwin Core format XML
}
\description{
dcw2XML - Convert Darwin Core format data frame to XML
}
\examples{
\dontrun{
 dcw2XML(dat)
} 
}

