% Generated by roxygen2 (4.0.0): do not edit by hand
\name{extract_flickrdb}
\alias{extract_flickrdb}
\title{extract_flickrdb - Extract Flickr data from sqlite database}
\usage{
extract_flickrdb(dbname = NA, outputfile = NA)
}
\arguments{
\item{dbname}{name of the database (Generated using function flickrtodatabase)}

\item{outputfile}{output csv file}
}
\value{
dataframe containing the Flickr data from database
}
\description{
extract_flickrdb - Extract Flickr data from sqlite database
}
\examples{
\dontrun{
mydat=extract_flickrdb("testdb","text.csv")
}
}

