% Generated by roxygen2 (4.0.0): do not edit by hand
\name{flickr2dwc}
\alias{flickr2dwc}
\title{flickr2dwc - Convert Flicker data frame to Darwin Core format}
\usage{
flickr2dwc(flickrdat = NA)
}
\arguments{
\item{flickrdat}{- Flickr Data frame to convert to Darwin Core}
}
\value{
data frame in Darwin Core format
}
\description{
flickr2dwc - Convert Flicker data frame to Darwin Core format
}
\examples{
\dontrun{
 flickr2dwc(flickrdat)
}
}

