% Generated by roxygen2 (4.0.0): do not edit by hand
\name{flickr2dwcA}
\alias{flickr2dwcA}
\title{flickr2dwcA - Convert Flicker data frame to Darwin Core Archiev file}
\usage{
flickr2dwcA(flickrdat = NA, outfile = NA)
}
\arguments{
\item{flickrdat}{- Flickr Data frame to convert to Darwin Core Archiev file}

\item{outfile}{- Output file name without a .zip extension}
}
\description{
flickr2dwcA - Convert Flicker data frame to Darwin Core Archiev file
}
\examples{
\dontrun{
 flickr2dwcA(flickrdat)
}
}

