% Generated by roxygen2 (4.0.0): do not edit by hand
\name{flickrlist}
\alias{flickrlist}
\title{flickrlist - Get list of media items on Flickr for the search string.}
\usage{
flickrlist(apikey = NA, stext = NA)
}
\arguments{
\item{apikey}{- API key provided by Flickr.com website. (Refer http://www.flickr.com/services/api/misc.api_keys.html for more details.)}

\item{stext}{- Test to search, a speice nems of common name.}
}
\description{
flickrlist - Get list of media items on Flickr for the search string.
}
\examples{
\dontrun{
 flickrlist(myapikey,"Danaus chrysippus")
}
}

