% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickrtodatabase.R
\name{flickrtodatabase}
\alias{flickrtodatabase}
\title{flickrtodatabase - Get Flickr data for a list of items and store into a database}
\usage{
flickrtodatabase(apikey, inputfile, inputfield, outdbname, dbfolder = ".",
  StIndex = 1)
}
\arguments{
\item{apikey}{- API key provided by Flickr.com website. (Refer http://www.flickr.com/services/api/misc.api_keys.html for more details.)}

\item{inputfile}{- Input csv file containing the list of species to search for.}

\item{inputfield}{- Fieldname containing the names to search.}

\item{outdbname}{- Name of the database}

\item{dbfolder}{- Folder in which the database should be stored.}

\item{StIndex}{- Starting Index. If is 1, then creates the databse file, otherwise appends the data in the database.}
}
\description{
flickrtodatabase - Get Flickr data for a list of items and store into a database
}
\examples{
\dontrun{
 flickrtodatabase(myapikey,"test.txt","scname","testdb")
}
}

