% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dcwToXML}
\alias{dcwToXML}
\title{dcwToXML - Convert Darwin Core format data frame to XML}
\usage{
dcwToXML(dat = NA)
}
\arguments{
\item{dat}{- Darwin Core Data frame to convert to XML}
}
\value{
Simple Darwin Core format XML
}
\description{
dcwToXML - Convert Darwin Core format data frame to XML
}
\examples{
\dontrun{
 dcwToXML(dat)
}
}

