% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flickrlist.R
\name{flickrlist}
\alias{flickrlist}
\title{flickrlist - Get list of media items on Flickr for the search string.}
\usage{
flickrlist(apikey = NA, stext = NA, bbox = NA)
}
\arguments{
\item{apikey}{- API key provided by Flickr.com website. (Refer http://www.flickr.com/services/api/misc.api_keys.html for more details.)}

\item{stext}{- Test to search, a species name of common name.}

\item{bbox}{- Bounding box. Geographical coordinates of Left top and right bottom corners separate by comma in string format i.e. "-180,-90,180,90" for whole world.

Using bbox parameter would not return photo without geocoding. If photos returned are more than 4000 the function automatically defaults to geocoded photos.}
}
\description{
flickrlist - Get list of media items on Flickr for the search string.
}
\examples{
\dontrun{
 flickrlist(myapikey,"Danaus chrysippus")
}
}
